gk: add functionalities to edit LPM table
This patch mainly adds support for addition/deletion of FIB entries
for the global GK LPM table, e.g., add_fib_entry() for addition, and
del_fib_entry() for deletion.

Actually, there is a hierarchical relationship among different types
of FIB entries. Generally, FIB entry for destination prefix -> FIB entry
for a Grantor -> FIB entry for a Gateway -> FIB entry for a Neighbor,
where "->" means a reference. However, there may be some special cases,
e.g., a Grantor is possibly a neighbor.

For the deletion of a FIB entry, before we remove it, we need to nofity
all the GK blocks to flush the flow entries that has a reference to this
FIB entry. After all the GK blocks complete this flush process, then we
can safely remove the FIB entry.